+++ src/webpackServer.js
@@ -1,19 +1,20 @@
+import createServerWebpackConfig from './createServerWebpackConfig'
 import debug from './debug'
 import devServer from './devServer'
-import createServerWebpackConfig from './createServerWebpackConfig'
+import {deepToString} from './utils'
 
 /**
  * Start a development server with Webpack using a given build configuration.
  */
 export default function(args, buildConfig, cb) {
-  // Force environment variable to development
+  // Force environment to development
   process.env.NODE_ENV = 'development'
 
   let {server = {staticPath: null}} = buildConfig
 
   let webpackConfig = createServerWebpackConfig(args, buildConfig)
 
-  debug('webpack config: %o', webpackConfig)
+  debug('webpack config: %s', deepToString(webpackConfig))
 
   devServer(webpackConfig, {
     fallback: !!args.fallback,